+++ core/liquidity/snapshot.go
@@ -31,6 +31,7 @@ import (
 type OrderBook interface {
 	GetOrderByID(orderID string) (*types.Order, error)
 	GetOrdersPerParty(party string) []*types.Order
+	GetLiquidityOrders(party string) []*types.Order
 }
 
 type SnapshotEngine struct {
@@ -39,22 +40,19 @@ type SnapshotEngine struct {
 	market string
 
 	// liquidity types
-	stopped                          bool
-	serialisedParameters             []byte
-	serialisedPartiesLiquidityOrders []byte
-	serialisedPartiesOrders          []byte
-	serialisedPendingProvisions      []byte
-	serialisedProvisions             []byte
-	serialisedSupplied               []byte
+	stopped                     bool
+	serialisedParameters        []byte
+	serialisedPendingProvisions []byte
+	serialisedProvisions        []byte
+	serialisedSupplied          []byte
 
 	// keys, need to be computed when the engine is
 	// instantiated as they are dynamic
-	hashKeys                  []string
-	parametersKey             string
-	partiesLiquidityOrdersKey string
-	pendingProvisionsKey      string
-	provisionsKey             string
-	suppliedKey               string
+	hashKeys             []string
+	parametersKey        string
+	pendingProvisionsKey string
+	provisionsKey        string
+	suppliedKey          string
 }
 
 func NewSnapshotEngine(config Config,
@@ -85,43 +83,6 @@ func NewSnapshotEngine(config Config,
 	se.buildHashKeys(market)
 
 	return se
-}
-
-func reconcileOrders(log *logging.Logger, orders *SnapshotablePartiesOrders, orderbook OrderBook) *SnapshotablePartiesOrders {
-	newOrders := newSnapshotablePartiesOrders()
-	for party, v := range orders.m {
-		newOrders.ResetForParty(party)
-		for _, o := range v {
-			order, err := orderbook.GetOrderByID(o.ID)
-			if err != nil {
-				// if not in the order book we just add the original
-				newOrders.Add(o.Party, o)
-				continue
-			}
-
-			if order.Status != o.Status ||
-				order.Remaining != o.Remaining ||
-				order.Price.NEQ(o.Price) {
-				log.Panic("snapshot contains unexpected mismatch between orderbook and lp-engine's orders",
-					logging.String("ob-status", order.Status.String()),
-					logging.String("lp-status", o.Status.String()),
-					logging.String("ob-price", order.Price.String()),
-					logging.String("lp-price", o.Price.String()),
-					logging.Uint64("ob-remaining", order.Remaining),
-					logging.Uint64("lp-remaining", o.Remaining),
-				)
-			}
-
-			newOrders.Add(order.Party, order)
-		}
-	}
-	return newOrders
-}
-
-// ReconcileWithOrderBook ensures that when restoring state from a snapshot the orders in the matching engine and
-// liquidity engine are pointers to the same underlying order struct.
-func (e *SnapshotEngine) ReconcileWithOrderBook(orderbook OrderBook) {
-	e.liquidityOrders = reconcileOrders(e.log, e.liquidityOrders, orderbook)
 }
 
 func (e *SnapshotEngine) UpdateMarketConfig(model risk.Model, monitor PriceMonitor) {
@@ -176,9 +137,6 @@ func (e *SnapshotEngine) LoadState(ctx context.Context, p *types.Payload) ([]typ
 			ctx, pl.Provisions.GetLiquidityProvisions(), p)
 	case *types.PayloadLiquidityParameters:
 		return nil, e.loadParameters(ctx, pl.Parameters, p)
-	case *types.PayloadLiquidityPartiesLiquidityOrders:
-		return nil, e.loadPartiesLiquidityOrders(
-			ctx, pl.PartiesLiquidityOrders.PartyOrders, p)
 	case *types.PayloadLiquiditySupplied:
 		return nil, e.loadSupplied(pl.LiquiditySupplied, p)
 	default:
@@ -192,25 +150,6 @@ func (e *SnapshotEngine) loadSupplied(ls *snapshotpb.LiquiditySupplied, p *types
 		return err
 	}
 	e.serialisedSupplied, err = proto.Marshal(p.IntoProto())
-	return err
-}
-
-func (e *SnapshotEngine) loadPartiesLiquidityOrders(
-	_ context.Context, pOrders []*snapshotpb.PartyOrders, p *types.Payload,
-) error {
-	e.Engine.liquidityOrders = newSnapshotablePartiesOrders()
-	for _, partyOrders := range pOrders {
-		e.Engine.liquidityOrders.ResetForParty(partyOrders.Party)
-		for _, v := range partyOrders.Orders {
-			order, err := types.OrderFromProto(v)
-			if err != nil {
-				return err
-			}
-			e.Engine.liquidityOrders.Add(order.Party, order)
-		}
-	}
-	var err error
-	e.serialisedPartiesLiquidityOrders, err = proto.Marshal(p.IntoProto())
 	return err
 }
 
@@ -281,8 +220,6 @@ func (e *SnapshotEngine) serialise(k string) ([]byte, error) {
 	switch k {
 	case e.parametersKey:
 		buf, changed, err = e.serialiseParameters()
-	case e.partiesLiquidityOrdersKey:
-		buf, changed, err = e.serialisePartiesLiquidityOrders()
 	case e.pendingProvisionsKey:
 		buf, changed, err = e.serialisePendingProvisions()
 	case e.provisionsKey:
@@ -308,8 +245,6 @@ func (e *SnapshotEngine) serialise(k string) ([]byte, error) {
 	switch k {
 	case e.parametersKey:
 		e.serialisedParameters = buf
-	case e.partiesLiquidityOrdersKey:
-		e.serialisedPartiesLiquidityOrders = buf
 	case e.pendingProvisionsKey:
 		e.serialisedPendingProvisions = buf
 	case e.provisionsKey:
@@ -336,33 +271,6 @@ func (e *SnapshotEngine) serialiseParameters() ([]byte, bool, error) {
 	}
 
 	return e.marshalPayload(payload)
-}
-
-func (e *SnapshotEngine) serialisePartiesLiquidityOrders() ([]byte, bool, error) {
-	payload := &snapshotpb.Payload{
-		Data: &snapshotpb.Payload_LiquidityPartiesLiquidityOrders{
-			LiquidityPartiesLiquidityOrders: &snapshotpb.LiquidityPartiesLiquidityOrders{
-				MarketId:    e.market,
-				PartyOrders: partyOrdersToProto(e.Engine.liquidityOrders.m),
-			},
-		},
-	}
-	return e.marshalPayload(payload)
-}
-
-func partyOrdersToProto(m map[string]map[string]*types.Order) []*snapshotpb.PartyOrders {
-	pOrders := make([]*snapshotpb.PartyOrders, 0, len(m))
-	for party, orders := range m {
-		partyOrders := &snapshotpb.PartyOrders{Party: party}
-		partyOrders.Orders = []*typespb.Order{}
-		for _, order := range orders {
-			partyOrders.Orders = append(partyOrders.Orders, order.IntoProto())
-		}
-		sort.SliceStable(partyOrders.Orders, func(i, j int) bool { return partyOrders.Orders[i].Id < partyOrders.Orders[j].Id })
-		pOrders = append(pOrders, partyOrders)
-	}
-	sort.SliceStable(pOrders, func(i, j int) bool { return pOrders[i].Party < pOrders[j].Party })
-	return pOrders
 }
 
 func (e *SnapshotEngine) serialisePendingProvisions() ([]byte, bool, error) {
@@ -424,11 +332,6 @@ func (e *SnapshotEngine) buildHashKeys(market string) {
 			MarketId: market,
 		},
 	}).Key()
-	e.partiesLiquidityOrdersKey = (&types.PayloadLiquidityPartiesLiquidityOrders{
-		PartiesLiquidityOrders: &snapshotpb.LiquidityPartiesLiquidityOrders{
-			MarketId: market,
-		},
-	}).Key()
 	e.pendingProvisionsKey = (&types.PayloadLiquidityPendingProvisions{
 		PendingProvisions: &snapshotpb.LiquidityPendingProvisions{
 			MarketId: market,
@@ -446,6 +349,6 @@ func (e *SnapshotEngine) buildHashKeys(market string) {
 		},
 	}).Key()
 
-	e.hashKeys = append([]string{}, e.parametersKey, e.partiesLiquidityOrdersKey,
+	e.hashKeys = append([]string{}, e.parametersKey,
 		e.pendingProvisionsKey, e.provisionsKey, e.suppliedKey)
 }