+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - [](https://github.com/vegaprotocol/vega/issues/xxxx) -
 
 ### 🐛 Fixes
-- [](https://github.com/vegaprotocol/vega/issues/xxxx) -
+- [6563](https://github.com/vegaprotocol/vega/issues/6563) - Liquidity engine reads orders directly from the matching engine
 
 
 ## 0.61.0
@@ -53,7 +53,6 @@
 ### 🐛 Fixes
 - [6410](https://github.com/vegaprotocol/vega/issues/6410) - Add input validation for the `EstimateFee` endpoint.
 - [6556](https://github.com/vegaprotocol/vega/issues/6556) - Limit ledger entries filtering complexity and potential number of items.
-- [6563](https://github.com/vegaprotocol/vega/issues/6563) - Keep order in sync between the `OrderBook` and the liquidity engine when amending during an auction
 - [6539](https://github.com/vegaprotocol/vega/issues/6539) - Fix total fee calculation in estimate order
 - [6584](https://github.com/vegaprotocol/vega/issues/6584) - Simplify `ListBalanceChanges`, removing aggregation and forward filling for now
 - [6583](https://github.com/vegaprotocol/vega/issues/6583) - Cancel wallet connection request if no wallet